+++ source/config/index.js
@@ -17,9 +17,20 @@ export let make = () => confine(definition)
     @return { Proxy }
       Configuration proxy. 
 */
-export let assign = (...configs) => 
-    configs.reduce((m, c) => (is.func(c) ? c(m.config) : m.config = c || {}, m), make())
+export let assign = async (...configs) => 
+{
+    let data = make();
 
+    for (let config of configs)
+    {
+        if (is.func(config)) 
+            await config(data.config);
+        else
+            data.config = config;
+    }
+
+    return data;
+}
 
 export { default as defaults } from './defaults.js'
 export { default as required } from './required.js'