+++ source/errors.js
@@ -1,9 +1,16 @@
 
 export class AcidValidateError extends Error
 {
-    constructor(message)
+    constructor(reason, prop, value)
     {
-        super(message);
+        super();
+
+        this.reason = reason;
+        this.prop = prop;
+        this.value = value;
+
+        this.message = `${prop} ${reason}`;
+        if (value !== void 0) this.message += `; received ${JSON.stringify(value) || value}`;
     }
 }
 
@@ -11,6 +18,6 @@ export class AcidROOperationError extends Error
 {
     constructor(name, isPlain)
     {
-        super(`The ${name}() operation is not supported for restricted ${isPlain ? 'objecs' : 'arrays'}.`);
+        super(`The ${name}() operation is not supported for restricted ${isPlain ? 'objects' : 'arrays'}.`);
     }
 }